chore: Migrate to a goldmark extension
