+++ pkg/mark/mermaid/mermaid_test.go
@@ -1,9 +1,10 @@
-package mark
+package mermaid
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/kovetskiy/mark/pkg/mark/attachment"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,10 +13,10 @@ func TestExtractMermaidImage(t *testing.T) {
 		name     string
 		markdown []byte
 		scale    float64
-		want     Attachment
+		want     attachment.Attachment
 		wantErr  assert.ErrorAssertionFunc
 	}{
-		{"example", []byte("graph TD;\n A-->B;"), 1.0, Attachment{
+		{"example", []byte("graph TD;\n A-->B;"), 1.0, attachment.Attachment{
 			// This is only the PNG Magic Header
 			FileBytes: []byte{0x89, 0x50, 0x4e, 0x47, 0xd, 0xa, 0x1a, 0xa},
 			Filename:  "example.png",
@@ -30,7 +31,7 @@ func TestExtractMermaidImage(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := processMermaidLocally(tt.name, tt.markdown, tt.scale)
+			got, err := ProcessMermaidLocally(tt.name, tt.markdown, tt.scale)
 			if !tt.wantErr(t, err, fmt.Sprintf("processMermaidLocally(%v, %v)", tt.name, string(tt.markdown))) {
 				return
 			}
